requirements_api.txt
# Python 3.10

# Django and REST Framework
django==4.2.6
djangorestframework==3.14.0

# Voice cloning dependencies
torch --pre --index-url https://download.pytorch.org/whl/nightly/cu126
torchvision --pre --index-url https://download.pytorch.org/whl/nightly/cu126
torchaudio --pre --index-url https://download.pytorch.org/whl/nightly/cu126
scipy==1.13.1
librosa==0.10.2
huggingface-hub
munch==4.0.0
einops==0.8.0
pydub==0.25.1
descript-audio-codec==1.0.0
transformers==4.46.3
soundfile==0.12.1
sounddevice==0.5.0
modelscope==1.18.1
funasr==1.1.5
numpy==1.26.4
hydra-core==1.3.2
pyyaml
python-dotenv
matplotlib
